Move shaker into its own file to fix lint error
