+++ src/actions/longPress.ts
@@ -1,5 +1,4 @@
 import _ from 'lodash'
-import Dispatch from '../@types/Dispatch'
 import DragThoughtZone from '../@types/DragThoughtZone'
 import DropThoughtZone from '../@types/DropThoughtZone'
 import Path from '../@types/Path'
@@ -14,7 +13,8 @@ import hasMulticursor from '../selectors/hasMulticursor'
 import { registerActionMetadata } from '../util/actionMetadata.registry'
 import haptics from '../util/haptics'
 import head from '../util/head'
-import alert, { alertActionCreator } from './alert'
+import shaker from '../util/shaker'
+import alert from './alert'
 import { expandHoverDownActionCreator as expandHoverDown } from './expandHoverDown'
 import { expandOnHoverTopActionCreator as expandHoverUp } from './expandHoverUp'
 
@@ -124,58 +124,6 @@ const longPress = (state: State, payload: Payload) => {
 
   return state
 }
-
-/** A utility that shakes if hit too much. Returns a hit function that takes an id representing the source of the hit. If the number of unique ids exceeds the SHAKE_THRESHOLD within DEBOUNCE_SHAKING, trigger onShake. */
-const Shaker = <T>(
-  onShake: (
-    /** Extra data that is passed to onShake from the last call to hit. This is used to pass dispatch from the dragInProgressActionCreator since importing the app store causes a circular import. */
-    data: T,
-  ) => void,
-) => {
-  // track the different draggingThoughts values that are dispatched within a period of time
-  const DEBOUNCE_SHAKING = 100
-
-  // the number of unique Paths that must be hovered over within the BEBOUNCE_SHAKING period to a trigger a shake
-  const SHAKE_THRESHOLD = 6
-
-  let repeatedMax = 0
-  const repeatedIds = new Map<string, number>()
-
-  /** Resets the repeat counts. */
-  const reset = () => {
-    repeatedMax = 0
-    repeatedIds.clear()
-  }
-
-  /* Reset counters when shaking stops. */
-  const shaking = _.debounce(reset, DEBOUNCE_SHAKING)
-
-  return (data: T, id?: string) => {
-    // count repeated ids
-    if (id) {
-      const count = (repeatedIds.get(id) || 0) + 1
-      repeatedIds.set(id, count)
-      repeatedMax = Math.max(repeatedMax, count)
-    }
-
-    // check if we reached the shake threshold
-    if (repeatedMax >= SHAKE_THRESHOLD) {
-      onShake(data)
-      reset()
-    }
-
-    // start the timer on each hit
-    shaking()
-  }
-}
-
-// abort drag-and-drop on shake
-const shaker = Shaker((dispatch: Dispatch) => {
-  // force drag to abort
-  // we need to make all drop targets, visible drag behavior, and drop handlers short circuit when the drag has been aborted
-  // react-dnd does not allow programmatic cancellation of drag
-  dispatch([longPressActionCreator({ value: LongPressState.DragCanceled }), alertActionCreator('✗ Drag cancelled')])
-})
 
 /** Action-creator for longPress. */
 export const longPressActionCreator =
